refactor: modularize 04-Bio-Follows-Alerts
