+++ src/Apps/ArtAdvisor/04-Bio-Follows-Alerts/App.tsx
@@ -8,6 +8,7 @@ import {
   StackableBorderBox,
   Box,
   Spinner,
+  useToasts,
 } from "@artsy/palette"
 import { useSystemContext } from "System/SystemContext"
 import Markdown from "marked-react"
